Add: Test that checks TwitterProducerTest sendMessage(String message) method
Checks that a record can successfully sent and acknowledged by a Kafka Docker Container.

Does not verify formatting of message, and verifies that record is of the correct offset value. Also this is not an integration test with actual Kafka cluster that will be used to run program in production